requirements.txt
streamlit
openai
PyMuPDF
python-dotenv